refactor: :art: cambiado el nombre e imports de Switch
